+++ wse_server/__init__.py
@@ -2,5 +2,5 @@
 
 from .router import SnapshotProvider, WSEConfig, create_wse_router
 
-__version__ = "1.3.6"
+__version__ = "1.3.7"
 __all__ = ["create_wse_router", "WSEConfig", "SnapshotProvider"]